Swap the firmware internal imu filter and the true pitch filter
Use the firmware internal filter as the "true pitch" filter, with a
"standard" Mahony setup. Change the Float package "true pitch" filter to
act as the main balancing filter with the special Mahony kp ~= 2.

Adds a new configuration option Accelerometer Confidence Decay, as
that's a parameter of the balancing filter a user may want to tweak.

Also, to ensure some sort of troublefree backwards compatibility, the
package now checks on each re-configuration if the firmware internal
filter has a kp of > 1, which indicates it's been configured to be the
main balancing filter. In such case, it overwrites the values with
defaults that were used for the "true pitch" filter, to make life easier
for users and avoid hard-to-debug misconfigurations.